Migrate accounts to new schema

+++ src/index.ts
@@ -12,6 +12,7 @@ import dotenv from "dotenv";
 import prisma from "./database";
 import Leveling from "./leveling";
 import Logging from "./logging";
+import { runMigrations } from "./migration";
 
 dotenv.config();
 
@@ -25,6 +26,7 @@ const client = new Client({
 });
 
 const GUILD_ID = "811256944953262102";
+runMigrations();
 
 client.once(Events.ClientReady, async (client) => {
     console.log(`Ready! Logged in as ${client.user.tag}`);
@@ -79,10 +81,19 @@ client.on(Events.GuildMemberRemove, async (member) => {
     try {
         if (member.user.bot) return;
 
-        const user = await prisma.user.findUnique({ where: { discordId: member.user.id } });
-        if (user) {
+        const account = await prisma.account.findUnique({ 
+            where: { 
+                platform_platformId: {
+                    platform: "DISCORD",
+                    platformId: member.user.id
+                }
+            },
+            include: { user: true }
+        });
+
+        if (account && account.user) {
             await prisma.user.delete({
-                where: { discordId: member.id },
+                where: { id: account.user.id },
             });
 
             Logging.log(client.guilds.cache.get(GUILD_ID), `Deleted user ${member.user.displayName} from database because they left the server!`);
